Consider wireless when configuring wired ifaces
When there is only one wired interface, we always configure it as
'internal' interface.  This happens even when there are wireless
interfaces available.  This patch changes that behavior.  When at least
one wireless interface is available, we will configure the single wired
interface as 'external'.

Currently, during first boot, we don't have this case hitting because of
lack of proprietary-firmware-free wireless interfaces.  However in
future, this is a common case for FreedomBox devices to have one wired
interface and one wireless interface.

Signed-off-by: James Valleroy <jvalleroy@mailbox.org>